+++ cli/cli/CHANGELOG.md
@@ -6,7 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [1.17.2]
-no changes
+
+### Fixed
+
+- `beam project {new/add}` commands work if called from other directory than the one with Beamable config.
 
 ## [1.17.1]
 
@@ -25,7 +28,6 @@ no changes
 - `beam services deploy` will force build services to `linux/amd64` CPU architecture for usage on Beamable Cloud
 - Fix `beam open-mongo` returning wrong command to execute when storage is not running
 - Fix open-swagger case sensitivity problem
-- `beam project {new/add}` commands work if called from other directory than the one with Beamable config.
 
 ### Changed