+++ cli/cli/CHANGELOG.md
@@ -6,7 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [1.17.2]
-no changes
+
+### Fixed
+
+- `beam project {new/add}` commands work if called from other directory than the one with Beamable config.
 
 ## [1.17.1]